codecparser: add EpbReader to handle generic "emulation prevention byte" things
the logic can be shared with vc1, h264, h265, maybe even more codecs
