+++ codecparsers/Makefile.am
@@ -1,6 +1,7 @@
 libyami_codecparser_source_c = \
 	bitReader.cpp \
 	bitWriter.cpp \
+	EpbReader.cpp \
 	nalReader.cpp \
 	dboolhuff.c \
 	$(NULL)
